% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.mate.H.R
\name{allocate.mate.H}
\alias{allocate.mate.H}
\title{allocate.mate.H}
\usage{
allocate.mate.H(H, parents, max_F = 1, method = "min_F")
}
\arguments{
\item{H}{is square additive genetic relationship matrix with individual idenifiers as rownames and colnames.}

\item{parents}{data frame with the following columns (class in parentheses):
\itemize{
 \item{'ID' is the individual identifier (character).} 
 \item{'SEX' is the sex of the individual - 'M' or 'F', for male and female respectively (character).} 
 \item{'EBV' Estimated breeding value (EBV) of the individual - can not be NA if applying assortative mating as the method (numeric).} 
 \item{'N_AS_PARENT' The number of families the indivdiual is to contribute to (integer).} 
}}

\item{max_F}{is the maxiumum inbreeding coefficient allowed in the offspring of candidate parents (numeric between 0 and 1)}

\item{method}{either 'min_F' (to minimise the average inbreeding in offspring) or 'assorative' (to apply assortative mating) (character)}
}
\value{
'summary' is a data frame containing a summary of all possible families generated from matings between parental candidates: 
\itemize{
 \item{'SELECTED' N represents unselected families, Y represents selected families (i.e. mate allocations to be generated) and All represents all possible families.}           
 \item{'COUNT_FAMS' count of families.}
 \item{MEAN_EBV' mean of family EBVs.}
 \item{'SD_EBV' standard deviation of family EBVs.}
 \item{'MIN_EBV' minimum of family EBVs.} 
 \item{'MIN_EBV' maximum of family EBVs.}  
 \item{MEAN_EBV' mean of family inbreeding coefficients (F).}
 \item{'SD_EBV' standard deviation of family inbreeding coefficients (F).}
 \item{'MIN_EBV' minimum of family inbreeding coefficients (F).} 
 \item{'MIN_EBV' maximum of family inbreeding coefficients (F).}  
}

'all_families' is a data frame containing details of all possible families able to be generted from matings between parental candidates: 
\itemize{
 \item{'SIRE' parental candidate as sire.}           
 \item{'DAM' parental candidate as dam.} 
 \item{'F' inbreeding coefficient of family members (i.e. the 'family F').}       
 \item{'EBV' mean of parental EBVs (i.e. the 'family EBV').}
 \item{'SELECTED' N represents an unselected family (i.e. family is not to be generated), Y represents a selected family (i.e. mate allocated - family to be generated).}           
}

'optimal_families' is a data frame containing details of mate allocations (i.e.families to be generated from matings between parental candidates): 
\itemize{
 \item{'SIRE' parental candidate allocated as the sire.}           
 \item{'DAM' parental candidate allocated as the dam.} 
 \item{'F' inbreeding coefficient of family members (i.e. the 'family F').}       
 \item{'EBV' mean of parental EBVs (i.e. the 'family EBV').}          
}
}
\description{
This function generates a mating list for a set of candidate parents.  
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of candidate parents and a relationship matrix (H) including these candidate parents.
}
\examples{
#Retrieve example data
data(H)
data(parents)

output <- allocate.mate.H(H, parents, max_F = 0.08, method = "min_F") 
output$summary
head(output$all_families)
head(output$optimal_families)
}
